Split out the base command class into mixins
The base command class was getting pretty long, and I want to move some
logic from the help command into it, so it made sense to split it up
into a base class and some mixin classes.

Pylint is happy to comply with accessing unknown properties if the class
name ends in mixin.

Now that I know how to do this I should be able to apply this to the
database, at some point, because that is crazy long.
